requirements.txt
cryptography==39.0.1
mysql-connector-python==8.0.32
